% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Rhat.R
\name{calculate_Rhat}
\alias{calculate_Rhat}
\title{Calculate Gelman Diagnostics for Fitted RaCE NMA Models}
\usage{
calculate_Rhat(mcmc, names = NULL, level = 0.95, multivariate = FALSE)
}
\arguments{
\item{mcmc}{MCMC draws from the RaCE NMA model, in the form of the model output of the \code{mcmc_raceNMA} function.}

\item{names}{A vector of intervention names (optional)}

\item{level}{The \code{level} parameter from the \code{gelman.diag} function in the \code{coda} package. Defaults to 0.95.}

\item{multivariate}{The \code{multivariate} parameter from the \code{gelman.diag} function in the \code{coda} package. Defaults to FALSE.}
}
\value{
Gelman diagnostics for the inputted MCMC chains, in the format of the output of the \code{gelman.diag} function in the \code{coda} package.
}
\description{
This function applies MCMC outputs from the mcmc_raceNMA function to the gelman.diag function in the coda package.
}
\examples{
mcmc <- mcmc_raceNMA(mu_hat=c(0,0,1,1), s=c(.1,.1,.1,.1), seed=1)
calculate_Rhat(mcmc,names=paste0("Treatment ",1:4))
}
