% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clustermatrix.R
\name{create_clustermatrix}
\alias{create_clustermatrix}
\title{Create a posterior clustering matrix for the interventions based on RaCE NMA models}
\usage{
create_clustermatrix(mcmc, names = NULL, label_ranks = NULL)
}
\arguments{
\item{mcmc}{MCMC draws from the RaCE NMA model, in the form of the model output of the \code{mcmc_RCMVN} function.}

\item{names}{A vector of intervention names (optional)}

\item{label_ranks}{A vector containing rank levels for which posterior rank probabilities should be displayed within the clustering matrix. Defaults to \code{NULL}, indicating no probabilities are displayed as text.}
}
\value{
A ggplot of a posterior clustering matrix for the interventions.
}
\description{
This function creates a posterior clustering matrix for the interventions in the form of a ggplot.
}
\examples{
mcmc <- mcmc_RCMVN(ybar=c(0,0,1,1), s=c(.1,.1,.1,.1), mu0=0.5, sigma0=5, tau=0.5,chains=2,seed=1)
create_clustermatrix(mcmc,names=paste0("Treatment ",1:4),label_ranks=c(1,3))
}
