% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cumulativeranking.R
\name{create_cumulativeranking}
\alias{create_cumulativeranking}
\title{Create a cumulative ranking plot in RaCE NMA models}
\usage{
create_cumulativeranking(mcmc, names = NULL)
}
\arguments{
\item{mcmc}{MCMC draws from the RaCE NMA model, in the form of the model output of the \code{mcmc_RCMVN} function.}

\item{names}{A vector of intervention names (optional)}
}
\value{
A ggplot of a cumulative ranking plot.
}
\description{
This function creates a cumulative ranking plot in the form of a ggplot.
}
\examples{
mcmc <- mcmc_RCMVN(ybar=c(0,0,1,1), s=c(.1,.1,.1,.1), mu0=0.5, sigma0=5, tau=0.5,chains=2,seed=1)
create_cumulativeranking(mcmc,names=paste0("Treatment ",1:4))
}
