% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_violinplot.R
\name{create_violinplot}
\alias{create_violinplot}
\title{Create posterior violin plots for the mu parameters in RaCE NMA models}
\usage{
create_violinplot(mcmc, names = NULL)
}
\arguments{
\item{mcmc}{MCMC draws from the RaCE NMA model, in the form of the model output of the \code{mcmc_RCMVN} function.}

\item{names}{A vector of intervention names (optional)}
}
\value{
A ggplot of posterior violin plots for the mu parameters.
}
\description{
This function creates posterior violin plots for the mu parameters in the form of a ggplot.
}
\examples{
mcmc <- mcmc_RCMVN(ybar=c(0,0,1,1), s=c(.1,.1,.1,.1), mu0=0.5, sigma0=5, tau=0.5,chains=2,seed=1)
create_violinplot(mcmc,names=paste0("Treatment ",1:4))
}
