% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createtrace_K.R
\name{createtrace_K}
\alias{createtrace_K}
\title{Create trace plots for the K parameter in RaCE NMA models}
\usage{
createtrace_K(mcmc)
}
\arguments{
\item{mcmc}{MCMC draws from the RaCE NMA model, in the form of the model output of the \code{mcmc_RCMVN} function.}
}
\value{
A ggplot of trace plots for the K parameter.
}
\description{
This function creates trace plots for the K parameter in the form of a ggplot.
}
\examples{
mcmc <- mcmc_RCMVN(ybar=c(0,0,1,1), s=c(.1,.1,.1,.1), mu0=0.5, sigma0=5, tau=0.5,chains=2,seed=1)
createtrace_K(mcmc)
}
