% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulativeprobplot_ranks.R
\name{cumulativeprobplot_ranks}
\alias{cumulativeprobplot_ranks}
\title{Create a cumulative ranking plot in RaCE NMA models}
\usage{
cumulativeprobplot_ranks(data = NULL, mcmc = NULL, names = NULL)
}
\arguments{
\item{data}{A NxJ matrix of data, where N is the number of observations and J the number of treatments. This feature is designed to display results from a standard NMA study.}

\item{mcmc}{MCMC draws from the RaCE NMA model, in the form of the model output of the \code{mcmc_raceNMA} function.}

\item{names}{A vector of intervention names (optional)}
}
\value{
A ggplot of a cumulative ranking plot.
}
\description{
This function creates a cumulative ranking plot in the form of a ggplot.
}
\examples{
mcmc <- mcmc_raceNMA(mu_hat=c(0,0,1,1), s=c(.1,.1,.1,.1), seed=1)
cumulativeprobplot_ranks(mcmc=mcmc,names=paste0("Treatment ",1:4))
}
