% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_partition_correlation.R
\name{sample_partition_correlation}
\alias{sample_partition_correlation}
\title{Partition sampling in Bayesian RaCE-NMA models (internal use only)}
\usage{
sample_partition_correlation(
  mu_hat,
  J,
  nu,
  g,
  K,
  mu0,
  sigma0,
  cov,
  tau = tau,
  b_g = 0.5,
  d_g = 0.5,
  logdmvn = NULL
)
}
\arguments{
\item{mu_hat}{A vector of estimated average relative intervention effects based on a previous NMA. The jth entry is the effect of intervention j.}

\item{J}{A numeric indicating the total number of interventions being compared.}

\item{nu}{A vector indicating current values for nu in the Gibbs sampler.}

\item{g}{A vector indicating current values for g in the Gibbs sampler.}

\item{K}{A vector indicating current values for K in the Gibbs sampler.}

\item{mu0}{The hyperparameter mu0, usually specified as the grand mean of the average intervention effects.}

\item{sigma0}{The hyperparameter sigma_0, usually a large number as to be minimally informative.}

\item{cov}{A variance covariance matrix of relative intervention effects based on a previous NMA. The (i,j) entry is the covariane between intervention i and j's effects.}

\item{tau}{The standard deviation of the Metropolis Hastings proposal distribution.}

\item{b_g}{The probability of "birth"ing a new partition cluster, if possible. Default is 0.5.}

\item{d_g}{The probability of "death"ing an existing partition cluster, if possible. Default is 0.5.}

\item{logdmvn}{An optional function of \code{(x, mean)} returning the multivariate normal log-density with covariance \code{cov}, as built by \code{make_logdmvnorm}. Supplying it avoids re-factorising \code{cov} on every call. If \code{NULL}, one is built from \code{cov}.}
}
\value{
A list containing updated values for g, nu, and K.
}
\description{
This function implements a reversible jump MCMC procedure for updating the parameter partition in Bayesian Rank-Clustered Estimation for Network Meta-Analysis models in the case when we assume correlation among treatment effects. For internal use only.
}
\examples{
sample_partition_correlation(mu_hat=c(1,1,1),J=3,nu=1,g=c(1,1,1),K=1,mu0=1,
 sigma0=1,cov=diag(rep(.1,3)),tau=0.1)

}
\keyword{internal}
