% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_partition_normal.R
\name{sample_partition_normal}
\alias{sample_partition_normal}
\title{Partition sampling in Bayesian RaCE-NMA models (internal use only)}
\usage{
sample_partition_normal(
  ybar,
  J,
  nu,
  g,
  K,
  mu,
  sigma0,
  s,
  tau = tau,
  b_g = 0.5,
  d_g = 0.5
)
}
\arguments{
\item{ybar}{A vector of estimated average relative treatment effects based on a previous NMA. The jth entry is the effect of treatment j.}

\item{J}{A numeric indicating the total number of treatments being compared.}

\item{nu}{A vector indicating current values for nu in the Gibbs sampler.}

\item{g}{A vector indicating current values for g in the Gibbs sampler.}

\item{K}{A vector indicating current values for K in the Gibbs sampler.}

\item{mu}{The hyperparameter mu_0, usually specified as the grand mean of the average treatment effects.}

\item{sigma0}{The hyperparameter sigma_0, usually a large number as to be minimally informative.}

\item{s}{A vector of the assumed standard deviation of each treatment. The jth entry is the standard deviation of treatment j.}

\item{tau}{The standard deviation of the Metropolis Hastings proposal distribution.}

\item{b_g}{The probability of "birth"ing a new partition cluster, if possible. Default is 0.5.}

\item{d_g}{The probability of "death"ing an existing partition cluster, if possible. Default is 0.5.}
}
\value{
A list containing updated values for g, nu, and K.
}
\description{
This function implements a reversible jump MCMC procedure for updating the parameter partition in Bayesian Rank-Clustered Estimation for Network Meta-Analysis models. For internal use only.
}
