% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplot_mu.R
\name{traceplot_mu}
\alias{traceplot_mu}
\title{Create trace plots for the mu parameter in RaCE NMA models}
\usage{
traceplot_mu(mcmc, names = NULL)
}
\arguments{
\item{mcmc}{MCMC draws from the RaCE NMA model, in the form of the model output of the \code{mcmc_RCMVN} function.}

\item{names}{A vector of intervention names (optional)}
}
\value{
A ggplot of trace plots for the mu parameter.
}
\description{
This function creates trace plots for the mu parameter in the form of a ggplot.
}
\examples{
mcmc <- mcmc_raceNMA(mu_hat=c(0,0,1,1), s=c(.1,.1,.1,.1), seed=1)
traceplot_mu(mcmc,names=paste0("Treatment ",1:4))
}
