% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wang_posterior.R
\docType{data}
\name{wang_posterior}
\alias{wang_posterior}
\title{Posterior relative treatment effects for non-baseline treatments based on Wang et. al (2022)}
\format{
data.frame in which the (i,j) entry is the relative treatment effect of treatment j in posterior draw i.
}
\usage{
wang_posterior
}
\description{
The dataset includes relative treatment effects for 10 non-baseline treatments across 60k posterior
draws based on an NMA study by Wang et. al (2022).
}
\keyword{datasets}
