% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_SUCRA_MNBT.R
\name{calculate_SUCRA_MNBT}
\alias{calculate_SUCRA_MNBT}
\title{Calculate SUCRA and MNBT based on MCMC draws in NMA or RaCE-NMA models}
\usage{
calculate_SUCRA_MNBT(data = NULL, mcmc = NULL, confidence = 0.5, names = NULL)
}
\arguments{
\item{data}{A NxJ matrix of data to display as a forest plot, where N is the number of observations and J the number of treatments. This feature is designed for use to display a forest plot of results from a standard NMA study.}

\item{mcmc}{MCMC draws from the RaCE NMA model, in the form of the model output of the \code{mcmc_RCMVN} function.}

\item{confidence}{The \code{confidence} parameter from the \code{gelman.diag} function in the \code{coda} package. Defaults to 0.95.}

\item{names}{A vector of intervention names (optional)}
}
\value{
A table containing SUCRA and MNBT values for each treatment, ordered by descending SUCRA values.
}
\description{
This function creates a results table of SUCRA and Median Number of Better Treatments (MNBT) based on
MCMC draws from a standard NMA model or a fitted RaCE-NMA model.
}
\examples{
mcmc <- mcmc_RCMVN(ybar=c(0,0,1,1), s=c(.1,.1,.1,.1), mu0=0.5, sigma0=5, tau=0.5,chains=2,seed=1)
calculate_SUCRA_MNBT(mcmc=mcmc)
}
