% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterplot_ranks.R
\name{clusterplot_ranks}
\alias{clusterplot_ranks}
\title{Create a posterior clustering matrix for the interventions based on RaCE NMA models}
\usage{
clusterplot_ranks(data = NULL, mcmc = NULL, names = NULL, label_ranks = NULL)
}
\arguments{
\item{data}{A NxJ matrix of data, where N is the number of observations and J the number of treatments. This feature is designed to display results from a standard NMA study.}

\item{mcmc}{MCMC draws from the RaCE NMA model, in the form of the model output of the \code{mcmc_RCMVN} function.}

\item{names}{A vector of intervention names (optional)}

\item{label_ranks}{A vector containing rank levels for which posterior rank probabilities should be displayed within the clustering matrix. Only non-zero probabilities are displayed. Defaults to \code{NULL}, indicating no probabilities are displayed as text.}
}
\value{
A ggplot of a posterior clustering matrix for the interventions.
}
\description{
This function creates a posterior clustering matrix for the interventions in the form of a ggplot.
}
\examples{
data("wang_posterior")
mcmc <- mcmc_raceNMA(posterior=wang_posterior,iter=2000)
clusterplot_ranks(mcmc=mcmc,names=paste0("Treatment ",1:10),label_ranks=1:3)
}
