% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clustermatrix.R
\name{create_clustermatrix}
\alias{create_clustermatrix}
\title{Create a posterior clustering matrix for the interventions based on RaCE NMA models}
\usage{
create_clustermatrix(
  data = NULL,
  mcmc = NULL,
  names = NULL,
  label_ranks = NULL
)
}
\arguments{
\item{data}{A NxJ matrix of data, where N is the number of observations and J the number of treatments. This feature is designed to display results from a standard NMA study.}

\item{mcmc}{MCMC draws from the RaCE NMA model, in the form of the model output of the \code{mcmc_RCMVN} function.}

\item{names}{A vector of intervention names (optional)}

\item{label_ranks}{A vector containing rank levels for which posterior rank probabilities should be displayed within the clustering matrix. Defaults to \code{NULL}, indicating no probabilities are displayed as text.}
}
\value{
A ggplot of a posterior clustering matrix for the interventions.
}
\description{
This function creates a posterior clustering matrix for the interventions in the form of a ggplot.
}
\examples{
data("wang_posterior")
mcmc <- mcmc_RCMVN(posterior=wang_posterior,num_iters=100)
create_clustermatrix(mcmc=mcmc,names=paste0("Treatment ",1:4),label_ranks=c(1,3))
}
