% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Toy data of posterior relative treatment effects}
\format{
data.frame in which the (i,j) entry is the relative treatment effect of treatment j in posterior draw i.
}
\usage{
toy_data
}
\description{
The dataset includes relative treatment effects for 4 treatments across 10k posterior
draws, as studied in Simulation Study 2 of Pearce and Zhou (2025).
}
\keyword{datasets}
