% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplot_K.R
\name{traceplot_K}
\alias{traceplot_K}
\title{Create trace plots for the K parameter in RaCE NMA models}
\usage{
traceplot_K(mcmc)
}
\arguments{
\item{mcmc}{MCMC draws from the RaCE NMA model, in the form of the model output of the \code{mcmc_RCMVN} function.}
}
\value{
A ggplot of trace plots for the K parameter.
}
\description{
This function creates trace plots for the K parameter in the form of a ggplot.
}
\examples{
mcmc <- mcmc_raceNMA(mu_hat=c(0,0,1,1), s=c(.1,.1,.1,.1), seed=1)
traceplot_K(mcmc)
}
